feat: leverage just dependency resolution
This reworks the justfile to accomplish 2 major goals.

First, remove internal `just` calls. This prevents embedding titanoboa
easily as submodule into other projects since your root directory will
traverse into the titanoboa. While this can be worked around, this
should mean that titanoboa could be used as a `just` submodule as part
of a `git` submodule

Second. This enables building using locally available packages. This is
for the squash and iso steps. The squash step could be run within a
rootful container nested under a rootless container; however, the iso
step cannot easily be ran in a nested configuration. This instead checks
if the required RPMs are installed already. If so, it will then run it
locally instead of in a container.

Finally, this removed all instances of internal `sudo` calls. Instead
invoke the build process with `sudo`.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>